voice-agent/requirements.txt
livekit-agents[deepgram,google,openai,elevenlabs,silero,sarvam]~=1.1.5
livekit-plugins-noise-cancellation~=0.2
python-dotenv
pyyaml

markitdown[all]

qdrant-client

llama-index
llama-index-vector-stores-qdrant
llama-index-embeddings-google-genai
llama-index-llms-google-genai
llama-index-readers-file
llama-index-embeddings-huggingface

rich~=14.0.0

litellm[proxy]

noisereduce~=3.0.3

livekit
llama-index-core
uvloop
httpx
requests
attrs
protobuf
sympy
fastapi

sarvamai
openai
uvicorn

langfuse>=2,<3

fastembed

google-genai
google-cloud-storage

pyee

opik